+++ baybe/kernels/base.py
@@ -125,7 +125,7 @@ class CompositeKernel(Kernel, ABC):
     """Abstract base class for all composite kernels."""
 
 
-# Register de-/serialization hooks
+# Register (un-)structure hooks
 converter.register_structure_hook(Kernel, get_base_structure_hook(Kernel))
 converter.register_unstructure_hook(Kernel, unstructure_base)