+++ baybe/parameters/substance.py
@@ -13,13 +13,16 @@ from baybe.parameters.base import DiscreteParameter
 from baybe.parameters.enum import SubstanceEncoding
 from baybe.parameters.validation import validate_decorrelation
 from baybe.utils.basic import group_duplicate_values
-from baybe.utils.dataframe import df_drop_single_value_columns, df_uncorrelated_features
+from baybe.utils.dataframe import (
+    add_noise_to_perturb_degenerate_rows,
+    df_drop_single_value_columns,
+    df_uncorrelated_features,
+)
 
 try:  # For python < 3.11, use the exceptiongroup backport
     ExceptionGroup
 except NameError:
     from exceptiongroup import ExceptionGroup
-
 
 Smiles = str
 """Type alias for SMILES strings."""
@@ -64,6 +67,14 @@ class SubstanceParameter(DiscreteParameter):
         default=SubstanceEncoding.MORDRED, converter=SubstanceEncoding
     )
     # See base class.
+
+    kwargs_fingerprint: dict[str, Any] = field(
+        factory=dict, validator=instance_of(dict)
+    )
+    """Keyword arguments passed to fingerprint generator."""
+
+    kwargs_conformer: dict[str, Any] = field(factory=dict, validator=instance_of(dict))
+    """Keyword arguments passed to conformer generator."""
 
     @data.validator
     def _validate_substance_data(  # noqa: DOC101, DOC103
@@ -118,27 +129,20 @@ class SubstanceParameter(DiscreteParameter):
         from baybe.utils import chemistry
 
         vals = list(self.data.values())
-        pref = self.name + "_"
+        pref = self.name
 
         # Get the raw descriptors
-        if self.encoding is SubstanceEncoding.MORDRED:
-            comp_df = chemistry.smiles_to_mordred_features(vals, prefix=pref)
-        elif self.encoding is SubstanceEncoding.RDKIT:
-            comp_df = chemistry.smiles_to_rdkit_features(vals, prefix=pref)
-        elif self.encoding is SubstanceEncoding.MORGAN_FP:
-            comp_df = chemistry.smiles_to_fp_features(vals, prefix=pref)
-        else:
-            raise ValueError(
-                f"Unknown parameter encoding {self.encoding} for parameter {self.name}."
-            )
+        comp_df = chemistry.smiles_to_fingerprint_features(
+            vals,
+            encoding=self.encoding,
+            prefix=pref,
+            kwargs_conformer=self.kwargs_conformer,
+            kwargs_fingerprint=self.kwargs_fingerprint,
+        )
 
         # Drop NaN and constant columns
         comp_df = comp_df.loc[:, ~comp_df.isna().any(axis=0)]
         comp_df = df_drop_single_value_columns(comp_df)
-
-        # If there are bool columns, convert them to int (possible for Mordred)
-        bool_cols = comp_df.select_dtypes(bool).columns
-        comp_df[bool_cols] = comp_df[bool_cols].astype(int)
 
         # Label the rows with the molecule names
         comp_df.index = pd.Index(self.values)
@@ -149,6 +153,9 @@ class SubstanceParameter(DiscreteParameter):
                 comp_df = df_uncorrelated_features(comp_df)
             else:
                 comp_df = df_uncorrelated_features(comp_df, threshold=self.decorrelate)
+
+        # Add noise to degenerate rows if present
+        add_noise_to_perturb_degenerate_rows(comp_df)
 
         return comp_df