+++ baybe/searchspace/continuous.py
@@ -37,6 +37,7 @@ from baybe.searchspace.validation import (
 from baybe.serialization import SerialMixin, converter, select_constructor_hook
 from baybe.utils.basic import to_tuple
 from baybe.utils.dataframe import get_transform_objects, pretty_print_df
+from baybe.utils.numerical import DTypeFloatNumpy
 from baybe.utils.plotting import to_string
 
 if TYPE_CHECKING:
@@ -95,7 +96,9 @@ class SubspaceContinuous(SerialMixin):
         nonlinear_df = pd.DataFrame(nonlin_constraints_list)
 
         fields = [
-            to_string("Continuous Parameters", pretty_print_df(param_df)),
+            to_string(
+                "Continuous Parameters", pretty_print_df(param_df, max_colwidth=None)
+            ),
             to_string("Linear Equality Constraints", pretty_print_df(lin_eq_df)),
             to_string("Linear Inequality Constraints", pretty_print_df(lin_ineq_df)),
             to_string("Non-linear Constraints", pretty_print_df(nonlinear_df)),
@@ -313,6 +316,7 @@ class SubspaceContinuous(SerialMixin):
         return pd.DataFrame(
             {p.name: p.bounds.to_tuple() for p in self.parameters},
             index=["min", "max"],
+            dtype=DTypeFloatNumpy,
         )
 
     def _drop_parameters(self, parameter_names: Collection[str]) -> SubspaceContinuous:
@@ -327,10 +331,14 @@ class SubspaceContinuous(SerialMixin):
         return SubspaceContinuous(
             parameters=[p for p in self.parameters if p.name not in parameter_names],
             constraints_lin_eq=[
-                c._drop_parameters(parameter_names) for c in self.constraints_lin_eq
+                c._drop_parameters(parameter_names)
+                for c in self.constraints_lin_eq
+                if set(c.parameters) - set(parameter_names)
             ],
             constraints_lin_ineq=[
-                c._drop_parameters(parameter_names) for c in self.constraints_lin_ineq
+                c._drop_parameters(parameter_names)
+                for c in self.constraints_lin_ineq
+                if set(c.parameters) - set(parameter_names)
             ],
         )