+++ baybe/targets/numerical.py
@@ -113,11 +113,18 @@ class NumericalTarget(Target, SerialMixin):
     def _validate_transformation(  # noqa: DOC101, DOC103
         self, _: Any, value: TargetTransformation | None
     ) -> None:
-        """Validate that the given transformation is compatible with the specified mode.
+        """Validate compatability between transformation, bounds and the mode.
 
         Raises:
+            ValueError: If a target transformation was provided for an unbounded
+                target.
             ValueError: If the target transformation and mode are not compatible.
         """
+        if (value is not None) and (not self.bounds.is_bounded):
+            raise ValueError(
+                f"You specified a transformation for target '{self.name}', but "
+                f"did not specify any bounds."
+            )
         if (value is not None) and (value not in _VALID_TRANSFORMATIONS[self.mode]):
             raise ValueError(
                 f"You specified bounds for target '{self.name}', but your "