+++ benchmarks/result/__init__.py
@@ -1,0 +1,6 @@
+"""Benchmark results."""
+
+from benchmarks.result.metadata import ResultMetadata
+from benchmarks.result.result import Result
+
+__all__ = ["Result", "ResultMetadata"]