+++ examples/Basics/recommenders.py
@@ -79,14 +79,14 @@ available_acq_functions = [
     "UCB",  # upper confidence bound with beta of 1.0
 ]
 
-# Note that the qvailability of the acquisition functions might depend on the `batch_size`:
+# Note that the availability of the acquisition functions might depend on the `batch_size`:
 #   - If `batch_size` is set to 1, all available acquisition functions can be chosen
 #   - If a larger value is chosen, only those that allow batching.
 #       That is, 'q'-variants of the acquisition functions must be chosen.
 
-# The default he acquisition function is q-Expected Improvement.
+# The default acquisition function is q-Log Expected Improvement.
 
-ACQ_FUNCTION = "qEI"
+ACQ_FUNCTION = "qLogEI"
 
 ### Other parameters