+++ tests/test_substance_parameter.py
@@ -19,3 +19,24 @@ from .conftest import run_iterations
 def test_run_iterations(campaign, batch_size, n_iterations):
     """Test running some iterations with fake results and a substance parameter."""
     run_iterations(campaign, n_iterations, batch_size)
+
+
+@pytest.mark.skipif(
+    not CHEM_INSTALLED, reason="Optional chem dependency not installed."
+)
+def test_degenerate_comp_df():
+    """Test a degenerate comp_df is detected and fixed numerically."""
+    from baybe.parameters import SubstanceParameter
+
+    # These molecules are know to cause a degenerate representation with rdkit fp's
+    dict_base = {
+        "Potassium acetate": r"O=C([O-])C.[K+]",
+        "Potassium pivalate": r"O=C([O-])C(C)(C)C.[K+]",
+        "Cesium acetate": r"O=C([O-])C.[Cs+]",
+        "Cesium pivalate": r"O=C([O-])C(C)(C)C.[Cs+]",
+    }
+    p = SubstanceParameter(name="p", data=dict_base, encoding="RDKITFINGERPRINT")
+
+    assert (
+        not p.comp_df.duplicated().any()
+    ), "A degenerate comp_df was not correctly treated."